Extract blackboard plugin crate
