+++ .github/workflows/docker-precheck.yml
@@ -69,6 +69,7 @@ jobs:
             crates/mesh-llm-guardrails/ \
             crates/mesh-llm-types/ \
             crates/mesh-llm-plugin/ \
+            crates/mesh-llm-plugin-blackboard/ \
             crates/mesh-client/ \
             crates/mesh-llm-api-client/ \
             crates/mesh-llm-api-server/ \
