+++ Cargo.toml
@@ -11,6 +11,7 @@ members = [
     "crates/mesh-llm-types",
     "crates/mesh-llm-ui",
     "crates/mesh-llm-plugin",
+    "crates/mesh-llm-plugin-blackboard",
     "crates/mesh-client",
     "crates/mesh-llm-api-client",
     "crates/mesh-llm-api-server",
