+++ Justfile
@@ -266,8 +266,12 @@ bundle output="/tmp/mesh-bundle.tar.gz":
     BUNDLE="$DIR/mesh-bundle"
     mkdir -p "$BUNDLE"
     cp {{ mesh_bin }} "$BUNDLE/"
+    for plugin in blackboard; do
+        bin="target/release/mesh-llm-plugin-$plugin"
+        [ -f "$bin" ] && cp "$bin" "$BUNDLE/"
+    done
     # Fix rpaths for portability
-    for bin in "$BUNDLE/mesh-llm"; do
+    for bin in "$BUNDLE"/*; do
         [ -f "$bin" ] || continue
         install_name_tool -add_rpath @executable_path/ "$bin" 2>/dev/null || true
     done
