+++ crates/mesh-llm-host-runtime/Cargo.toml
@@ -52,9 +52,9 @@ async-trait = "0.1"
 arboard = "3"
 rand = "0.10"
 base64 = "0.22"
-biip = "0.9.2"
 dirs = "6.0.0"
 hex = "0.4.3"
+json5 = "1.3.1"
 nostr-sdk = { version = "0.44.1", default-features = false }
 opentelemetry = { version = "0.31.0", default-features = false, features = ["metrics"] }
 opentelemetry_sdk = { version = "0.31.0", default-features = false, features = ["metrics"] }
@@ -88,10 +88,10 @@ toml = "0.9"
 zip = { version = "2", default-features = false, features = ["deflate"] }
 hf_hub = { package = "hf-hub", version = "1.0.0-rc.1", default-features = false, features = ["blocking"] }
 tabwriter = "1"
-json5 = "1.3.1"
 
 [dev-dependencies]
 axum = "0.8"
 serial_test = "3"
 mesh-client = { package = "mesh-llm-client", path = "../mesh-client" }
+mesh-llm-plugin-blackboard = { path = "../mesh-llm-plugin-blackboard" }
 tempfile = "3"
