+++ crates/mesh-llm-host-runtime/src/plugin/config.rs
@@ -1,7 +1,6 @@
 use super::{
-    BLACKBOARD_PLUGIN_ID, BLOBSTORE_PLUGIN_ID, CLAUDE_PLUGIN_ID, FLASH_MOE_PLUGIN_ID,
-    GOOSE_PLUGIN_ID, OPENAI_ENDPOINT_PLUGIN_ID, OPENCODE_PLUGIN_ID, PI_PLUGIN_ID, PluginSummary,
-    TELEMETRY_PLUGIN_ID,
+    BLACKBOARD_PLUGIN_ID, BLOBSTORE_PLUGIN_ID, FLASH_MOE_PLUGIN_ID, OPENAI_ENDPOINT_PLUGIN_ID,
+    PluginSummary, TELEMETRY_PLUGIN_ID,
 };
 use anyhow::{Context, Result, bail};
 use mesh_llm_plugin::MeshVisibility;
@@ -1996,19 +1995,10 @@ pub fn resolve_plugins(config: &MeshConfig, _host_mode: PluginHostMode) -> Resul
 }
 
 pub fn blackboard_plugin_spec() -> Result<ExternalPluginSpec> {
-    let command = std::env::current_exe()
-        .context("Cannot determine mesh-llm executable path")?
-        .display()
-        .to_string();
     Ok(ExternalPluginSpec {
         name: BLACKBOARD_PLUGIN_ID.to_string(),
-        command,
-        args: vec![
-            "--log-format".into(),
-            "json".into(),
-            "--plugin".into(),
-            BLACKBOARD_PLUGIN_ID.into(),
-        ],
+        command: bundled_plugin_command(BLACKBOARD_PLUGIN_ID),
+        args: Vec::new(),
         url: None,
         env: BTreeMap::new(),
     })
@@ -2148,33 +2138,20 @@ pub fn bundled_cli_plugin_spec(name: &str) -> Result<Option<ExternalPluginSpec>>
         return blackboard_plugin_spec().map(Some);
     }
 
-    if matches!(
-        name,
-        GOOSE_PLUGIN_ID | CLAUDE_PLUGIN_ID | PI_PLUGIN_ID | OPENCODE_PLUGIN_ID
-    ) {
-        return bundled_agent_cli_plugin_spec(name).map(Some);
-    }
-
     Ok(None)
 }
 
-fn bundled_agent_cli_plugin_spec(name: &str) -> Result<ExternalPluginSpec> {
-    let command = std::env::current_exe()
-        .context("Cannot determine mesh-llm executable path")?
-        .display()
-        .to_string();
-    Ok(ExternalPluginSpec {
-        name: name.to_string(),
-        command,
-        args: vec![
-            "--log-format".into(),
-            "json".into(),
-            "--plugin".into(),
-            name.to_string(),
-        ],
-        url: None,
-        env: BTreeMap::new(),
-    })
+fn bundled_plugin_command(name: &str) -> String {
+    let executable = format!("mesh-llm-plugin-{name}{}", std::env::consts::EXE_SUFFIX);
+    if let Ok(current_exe) = std::env::current_exe()
+        && let Some(parent) = current_exe.parent()
+    {
+        let sibling = parent.join(&executable);
+        if sibling.exists() {
+            return sibling.display().to_string();
+        }
+    }
+    executable
 }
 
 #[cfg(test)]
