+++ crates/mesh-llm-plugin-blackboard/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "mesh-llm-plugin-blackboard"
+version.workspace = true
+edition = "2024"
+
+[lints]
+workspace = true
+
+[dependencies]
+anyhow = "1"
+biip = "0.9.2"
+clap = { version = "4", features = ["derive"] }
+dirs = "6.0.0"
+mesh-llm-plugin = { path = "../mesh-llm-plugin" }
+regex-lite = "0.1"
+reqwest = { version = "0.12", features = ["json"] }
+rmcp = { version = "1.2", features = ["server", "transport-io"] }
+schemars = "1"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+tokio = { version = "1", features = ["full"] }
+
+[dev-dependencies]
+tempfile = "3"
