+++ crates/mesh-llm-plugin-blackboard/skills/blackboard/SKILL.md
