+++ crates/mesh-llm-plugin-blackboard/src/main.rs
@@ -1,0 +1,3 @@
+fn main() {
+    std::process::exit(mesh_llm_plugin_blackboard::run_main());
+}
