+++ crates/mesh-llm-plugin-blackboard/src/mcp.rs
