+++ docker/Dockerfile.client
@@ -50,10 +50,12 @@ COPY crates/mesh-llm/Cargo.toml crates/mesh-llm/Cargo.toml
 COPY crates/mesh-llm/build.rs crates/mesh-llm/build.rs
 COPY crates/mesh-llm-plugin/Cargo.toml crates/mesh-llm-plugin/Cargo.toml
 COPY crates/mesh-llm-plugin/build.rs crates/mesh-llm-plugin/build.rs
+COPY crates/mesh-llm-plugin-blackboard/Cargo.toml crates/mesh-llm-plugin-blackboard/Cargo.toml
 # Satisfy Cargo workspace member resolution
 COPY crates/mesh-llm-host-runtime/ crates/mesh-llm-host-runtime/
 COPY crates/mesh-llm/ crates/mesh-llm/
 COPY crates/mesh-llm-plugin/ crates/mesh-llm-plugin/
+COPY crates/mesh-llm-plugin-blackboard/ crates/mesh-llm-plugin-blackboard/
 COPY crates/mesh-client/ crates/mesh-client/
 COPY crates/mesh-llm-api-client/ crates/mesh-llm-api-client/
 COPY crates/mesh-llm-api-server/ crates/mesh-llm-api-server/
@@ -115,10 +117,12 @@ COPY crates/mesh-llm/Cargo.toml crates/mesh-llm/Cargo.toml
 COPY crates/mesh-llm/build.rs crates/mesh-llm/build.rs
 COPY crates/mesh-llm-plugin/Cargo.toml crates/mesh-llm-plugin/Cargo.toml
 COPY crates/mesh-llm-plugin/build.rs crates/mesh-llm-plugin/build.rs
+COPY crates/mesh-llm-plugin-blackboard/Cargo.toml crates/mesh-llm-plugin-blackboard/Cargo.toml
 # Satisfy Cargo workspace member resolution
 COPY crates/mesh-llm-host-runtime/ crates/mesh-llm-host-runtime/
 COPY crates/mesh-llm/ crates/mesh-llm/
 COPY crates/mesh-llm-plugin/ crates/mesh-llm-plugin/
+COPY crates/mesh-llm-plugin-blackboard/ crates/mesh-llm-plugin-blackboard/
 COPY crates/mesh-client/ crates/mesh-client/
 COPY crates/mesh-llm-api-client/ crates/mesh-llm-api-client/
 COPY crates/mesh-llm-api-server/ crates/mesh-llm-api-server/
