+++ scripts/build-linux.sh
@@ -240,23 +240,29 @@ elif [[ -d "$UI_DIR" ]]; then
 fi
 
 if [[ "${MESH_LLM_BUILD_PROFILE:-debug}" == "dev" || "${MESH_LLM_BUILD_PROFILE:-debug}" == "debug" ]]; then
-    echo "Building mesh-llm (profile: dev, bin only)..."
+    echo "Building mesh-llm and bundled CLI plugins (profile: dev)..."
     cargo_features=()
     case "$BACKEND" in
         cuda) cargo_features=(--features gpu-bench-cuda) ;;
         rocm) cargo_features=(--features gpu-bench-hip) ;;
     esac
-    (cd "$REPO_ROOT" && cargo build -p mesh-llm --bin mesh-llm "${cargo_features[@]}")
-    echo "Mesh binary: target/debug/mesh-llm"
+    (cd "$REPO_ROOT" && cargo build \
+        -p mesh-llm \
+        -p mesh-llm-plugin-blackboard \
+        "${cargo_features[@]}")
+    echo "Mesh binaries: target/debug/mesh-llm, target/debug/mesh-llm-plugin-blackboard"
 elif [[ "${MESH_LLM_BUILD_PROFILE:-debug}" == "release" ]]; then
-    echo "Building mesh-llm (profile: release)..."
+    echo "Building mesh-llm and bundled CLI plugins (profile: release)..."
     cargo_features=()
     case "$BACKEND" in
         cuda) cargo_features=(--features gpu-bench-cuda) ;;
         rocm) cargo_features=(--features gpu-bench-hip) ;;
     esac
-    (cd "$REPO_ROOT" && cargo build --release -p mesh-llm "${cargo_features[@]}")
-    echo "Mesh binary: target/release/mesh-llm"
+    (cd "$REPO_ROOT" && cargo build --release \
+        -p mesh-llm \
+        -p mesh-llm-plugin-blackboard \
+        "${cargo_features[@]}")
+    echo "Mesh binaries: target/release/mesh-llm, target/release/mesh-llm-plugin-blackboard"
 else
     echo "Unsupported MESH_LLM_BUILD_PROFILE '${MESH_LLM_BUILD_PROFILE:-debug}'. Expected debug, dev, or release." >&2
     exit 1
