+++ scripts/plan-clippy-batches.sh
@@ -22,6 +22,7 @@ WORKSPACE_MEMBERS=(
   "mesh-llm-types"
   "mesh-llm-ui"
   "mesh-llm-plugin"
+  "mesh-llm-plugin-blackboard"
   "mesh-llm-client"
   "mesh-llm-api-client"
   "mesh-llm-api-server"
